Change register to login/token (login/password earlier), delete selenium dependency

+++ modules/tgbot/handlers/form.py
@@ -7,13 +7,12 @@ from ..utils.logger import print_msg
 from ..utils.throttling import rate_limit
 from ..keyboards.moodle import add_delete_button
 
-from modules.functions.login import auth_microsoft, is_cookies_valid
 from ...functions.parser import Parser
 
 
 class RegForm(StatesGroup):
     waiting_for_barcode = State()
-    waiting_for_password = State()
+    waiting_for_token = State()
 
 
 @print_msg
@@ -38,21 +37,21 @@ async def wait_barcode(message: types.Message, state: FSMContext):
     if not barcode.isdigit():
         msg = await message.answer('Error\! Write your *barcode* one more time:', parse_mode='MarkdownV2')
     else:
-        msg = await message.answer("Write your *password*:", parse_mode='MarkdownV2')
-        await RegForm.waiting_for_password.set()
+        msg = await message.answer("Write your *token*:", parse_mode='MarkdownV2')
+        await RegForm.waiting_for_token.set()
 
     async with state.proxy() as data:
         data['barcode'] = barcode
         data['msg_del'] = msg
 
 
-async def wait_password(message: types.Message, state: FSMContext):
-    password = message.text
+async def wait_token(message: types.Message, state: FSMContext):
+    token = message.text
     async with state.proxy() as data:
         await delete_msg(data['msg_del'], message)
 
-    if len(password.split()) > 1:
-        msg = await message.answer('Error\! Write your *password* one more time:', parse_mode='MarkdownV2')
+    if len(token.split()) > 1:
+        msg = await message.answer('Error\! Write your *token* one more time:', parse_mode='MarkdownV2')
         async with state.proxy() as data:
             data['msg_del'] = msg
     else:
@@ -60,15 +59,15 @@ async def wait_password(message: types.Message, state: FSMContext):
         async with state.proxy() as data:
             barcode = data['barcode']
             data['msg_del'] = msg
-        cookies = await auth_microsoft(barcode, password)
-        if await is_cookies_valid(cookies):
             parser = Parser()
-            token, userid = await parser.get_token_and_userid(cookies)
-            await db.register_moodle_user(message.from_user.id, barcode, password, cookies, userid, token)
-            await message.answer("Your Moodle account is registred\!", parse_mode='MarkdownV2',
-                                 reply_markup=add_delete_button())
-        else:
-            await message.answer("Your barcode or password are invalid. Try again!", reply_markup=add_delete_button())
+        moodle_userid = await parser.get_user_id(token)
+        if moodle_userid is None:
+            await message.answer("Error\! Your *token* is invalid\!", parse_mode='MarkdownV2')
+            await state.finish()
+            return
+        await db.register_moodle_user(message.from_user.id, barcode, moodle_userid, token)
+        await message.answer("Your Moodle account is registred\!", parse_mode='MarkdownV2',
+                                reply_markup=add_delete_button())
         async with state.proxy() as data:
             await delete_msg(data['msg_del'])
         await state.finish()
@@ -86,4 +85,4 @@ async def delete_msg(*msgs: types.Message):
 def register_form(dp: Dispatcher):
     dp.register_message_handler(register_moodle, commands=['register'])
     dp.register_message_handler(wait_barcode, content_types=['text'], state=RegForm.waiting_for_barcode)
-    dp.register_message_handler(wait_password, content_types=['text'], state=RegForm.waiting_for_password)
+    dp.register_message_handler(wait_token, content_types=['text'], state=RegForm.waiting_for_token)
